franky fixes a whole bunch of warnings to compile in a more pedantic fashion also some compiler flags have been added to visual studio build so that we are stricter and safer

+++ lib/renderer/Sprite.cpp
@@ -142,14 +142,14 @@ void Sprite::Render() const
 	Real halfWidth = (mHeader->width/2.0)*mScale.x;
 	Real halfHeight = (mHeader->height/2.0)*mScale.y;
 
-	glTexCoord2f (bottomTextureCoord.x, topTextureCoord.y);
-	glVertex3f (halfWidth+mPosition.x, -halfHeight+mPosition.y, 0.0);
-	glTexCoord2f (topTextureCoord.x, topTextureCoord.y);
-	glVertex3f (-halfWidth+mPosition.x, -halfHeight+mPosition.y, 0.0);
-	glTexCoord2f (bottomTextureCoord.x, bottomTextureCoord.y);
-	glVertex3f (halfWidth+mPosition.x, halfHeight+mPosition.y, 0.0);
-	glTexCoord2f (topTextureCoord.x, bottomTextureCoord.y);
-	glVertex3f (-halfWidth+mPosition.x, halfHeight+mPosition.y, 0.0);
+	glTexCoord2f (GLfloat(bottomTextureCoord.x), GLfloat(topTextureCoord.y));
+	glVertex3f (GLfloat(halfWidth+mPosition.x), GLfloat(-halfHeight+mPosition.y), 0.0);
+	glTexCoord2f (GLfloat(topTextureCoord.x), GLfloat(topTextureCoord.y));
+	glVertex3f (GLfloat(-halfWidth+mPosition.x), GLfloat(-halfHeight+mPosition.y), 0.0);
+	glTexCoord2f (GLfloat(bottomTextureCoord.x), GLfloat(bottomTextureCoord.y));
+	glVertex3f (GLfloat(halfWidth+mPosition.x), GLfloat(halfHeight+mPosition.y), 0.0);
+	glTexCoord2f (GLfloat(topTextureCoord.x), GLfloat(bottomTextureCoord.y));
+	glVertex3f (GLfloat(-halfWidth+mPosition.x), GLfloat(halfHeight+mPosition.y), 0.0);
 #else
 	glTexCoord2f (0.0, 0.0);
 	glVertex3f (-5, -5, 0.0);
